Zoom out doesn't work correctly for simple surveys Fixes #6363
